refactor(site): extract sub-components from the behavior settings god component
Break AgentSettingsBehaviorPageView into focused sub-components,
each owning their own local state, derived state, and handlers.

- PersonalInstructionsSettings: user prompt textarea form
- SystemInstructionsSettings: admin system prompt + TextPreviewDialog
- VirtualDesktopSettings: admin desktop toggle
- WorkspaceAutostopSettings: admin autostop toggle + duration form
- RetentionPeriodSettings: admin retention toggle + number input

The isPromptSaving coupling is preserved via isAnyPromptSaving prop
passed to both prompt components.

Addresses review feedback from DanielleMaywood on PR #23833.
